#2892 fix cyclic dependencies
