+++ packages/plugin-ext/src/plugin/types-impl.ts
@@ -18,8 +18,8 @@ import { illegalArgument } from '../common/errors';
 import * as theia from '@theia/plugin';
 import URI from 'vscode-uri';
 import { relative } from '../common/paths-util';
-import { isMarkdownString } from './type-converters';
 import { startsWithIgnoreCase } from '../common/strings';
+import { MarkdownString, isMarkdownString } from './markdown-string';
 
 export class Disposable {
     private disposable: undefined | (() => void);
@@ -507,36 +507,6 @@ export class SnippetString {
         }
         this.value += '}';
 
-        return this;
-    }
-}
-
-export class MarkdownString {
-
-    value: string;
-    isTrusted?: boolean;
-
-    constructor(value?: string) {
-        this.value = value || '';
-    }
-
-    appendText(value: string): MarkdownString {
-        // escape markdown syntax tokens: http://daringfireball.net/projects/markdown/syntax#backslash
-        this.value += value.replace(/[\\`*_{}[\]()#+\-.!]/g, '\\$&');
-        return this;
-    }
-
-    appendMarkdown(value: string): MarkdownString {
-        this.value += value;
-        return this;
-    }
-
-    appendCodeblock(code: string, language: string = ''): MarkdownString {
-        this.value += '\n```';
-        this.value += language;
-        this.value += '\n';
-        this.value += code;
-        this.value += '\n```\n';
         return this;
     }
 }
